+++ lib/Web/Browser.hs
@@ -1,3 +1,5 @@
+{-# LANGUAGE LambdaCase #-}
+
 {-|
 Module      : Web.Browser
 Description : Open a web browser from Haskell
@@ -11,14 +13,18 @@ Linux and BSD.
 
 module Web.Browser
   ( openBrowser
+    -- * Utilities
+  , openBrowserWithExitCode
   ) where
 
+import           Control.Exception ( Exception (..), SomeException, try)
+import           System.Exit ( ExitCode (..) )
 import qualified Web.Browser.OS as OS
 
--- | Seeks to open the given item. If the item is a URL or another item
--- associated with a web browser (for example, it represents a local @.html@
--- file), seeks to open it in the user's preferred web browser. Returns whether
--- or not the operation succeeded.
+-- | Seeks to open the given item, silently. If the item is a URL or another
+-- item associated with a web browser (for example, it represents a local
+-- @.html@ file), seeks to open it in the user's preferred web browser. Returns
+-- whether or not the operation succeeded.
 --
 -- No checks are performed on the nature or validity of the given item.
 --
@@ -32,12 +38,32 @@ import qualified Web.Browser.OS as OS
 --   icon; and
 --
 -- * on Linux, FreeBSD, OpenBSD or NetBSD, the \'xdg-open\' script, if it
---   is on the user's PATH, via \'sh\' to allow the script's output to be
---   silenced.
+--   is on the user's PATH.
 --
 -- On other operating systems, the operation always fails.
+--
+-- @since 0.1.0.0
 openBrowser ::
      String
-     -- ^ URL or other item to try to open
+     -- ^ URL or other item to try to open.
   -> IO Bool
-openBrowser = OS.openBrowser
+openBrowser url = tryOpenUrl >>= \case
+  Left _ -> pure False
+  Right (ec, _, _) -> pure $ ec == ExitSuccess
+ where
+  tryOpenUrl :: IO (Either SomeException (ExitCode, String, String))
+  tryOpenUrl = openBrowserWithExitCode url
+
+-- | Exported to help with debugging. As for 'openBrowser' but returns either an
+-- exception or, as a triple, the 'ExitCode' of the opening mechanism and any
+-- output to the standard output and standard error channels. On failure, the
+-- meaning of the exit code will depend on the operating system; for unsupported
+-- operating systems, it will be 'ExitFailure' @1@.
+--
+-- @since UNRELEASED
+openBrowserWithExitCode ::
+     Exception e
+  => String
+     -- ^ URL or other item to try to open.
+  -> IO (Either e (ExitCode, String, String))
+openBrowserWithExitCode url = try $ OS.openBrowserWithExitCode url