+++ lib/unix-like/open/Web/Browser/OS.hs
@@ -4,14 +4,15 @@
 --------------------------------------------------------------------------------
 
 module Web.Browser.OS
-  ( openBrowser
+  ( openBrowserWithExitCode
   ) where
 
-import Utils ( openBrowserSilentlyWith )
+import System.Exit ( ExitCode (..) )
+import System.Process ( readProcessWithExitCode )
 
 -- https://ss64.com/mac/open.html
-openBrowser ::
+openBrowserWithExitCode ::
      String
-     -- ^ URL
-  -> IO Bool
-openBrowser url = openBrowserSilentlyWith "open" [url]
+     -- ^ URL or other item to try to open.
+  -> IO (ExitCode, String, String)
+openBrowserWithExitCode url = readProcessWithExitCode "open" [url] ""