+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to the
 
 * On Unix-like operating systems, use Haskell to silence the opening
   application or script, rather than the `sh` shell on Linux and BSD.
+* Add utility `openBrowserWithErrorCode`, to help with debugging failure.
 * Improve Haddock documentation.
 
 ## 0.3.0.1 - 2025-03-17